projetoAcademify/src/app/paginas/editar-estudantes/editar-estudantes.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { FormsModule } from '@angular/forms';
import { ActivatedRoute, Router } from '@angular/router';
import { EditarEstudantesComponent } from './editar-estudantes.component';

describe('EditarEstudantesComponent', () => {
  let component: EditarEstudantesComponent;
  let fixture: ComponentFixture<EditarEstudantesComponent>;
  let mockRouter: jasmine.SpyObj<Router>;

  beforeEach(async () => {
    mockRouter = jasmine.createSpyObj('Router', ['navigate']);
    await TestBed.configureTestingModule({
      declarations: [EditarEstudantesComponent],
      imports: [FormsModule],
      providers: [
        { provide: Router, useValue: mockRouter },
        { provide: ActivatedRoute, useValue: { snapshot: { paramMap: { get: () => '1' } } } }
      ]
    }).compileComponents();

    fixture = TestBed.createComponent(EditarEstudantesComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should populate student data on init', () => {
    expect(component.student).toEqual({ id: 1, name: 'João Silva', birthDate: '2000-01-01' });
  });

  it('should navigate to /estudantes on cancel', () => {
    component.onCancel();
    expect(mockRouter.navigate).toHaveBeenCalledWith(['/estudantes']);
  });
});
